Refactor package layout and remove unused modules
