+++ monitor_module/session.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from .ssh import SSHClient
+from device_clients.ssh import SSHClient
 from .config import Config
 from .utils import split_line
